fix(font_configurator): extract to separate package
