+++ font_configurator/test_assets/fontconfigs/valid_dummy__multiple_font_dirs.conf
