+++ font_configurator/test_assets/fontconfigs/valid_dummy__no_description.conf
